api/requirements.txt
requests
fastapi
uvicorn
openai
asyncio
python-dateutil
numpy